feat: restructure application to use serverless functions and modular routes netlify
