+++ .gitignore
@@ -1,2 +1,5 @@
 .env
 node_modules
+
+# Local Netlify folder
+.netlify
