+++ functions/hello.js
@@ -1,0 +1,6 @@
+exports.handler = async (event, context) => {
+    return {
+      statusCode: 200,
+      body: JSON.stringify({ message: "Hello, World!" }),
+    };
+  };
