+++ netlify.toml
@@ -1,0 +1,8 @@
+
+[build]
+    functions = "functions"
+
+[[redirects]]
+    from = "/*"
+    to = "/.netlify/functions/api"
+    status = 200
