+++ package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
+    "serverless-http": "^3.2.0",
     "twilio": "^5.4.4"
   }
 }
